requirements.txt
streamlit
google-generativeai
langchain
python-dotenv
pyPDF2
chromadb